Add a "Max seed money" option
This makes it so people with limited funds can put how much money
they're willing to spend on seeds, and we'll limit the number of
crops based on that (up to the value in "Number of crops").

This can be useful because some of the most profitable crops have
expensive seeds, so in the early game a large number of less profitable
crops can be better.